perf: 菜单按钮组件化
