+++ src/components/pptist/Editor/EditorHeader/index.vue
@@ -1,28 +1,7 @@
 <template>
   <div class="editor-header">
     <div class="left">
-      <Popover trigger="click" placement="bottom-start" v-model:value="mainMenuVisible">
-        <template #content>
-          <FileInput accept=".pptist"  @change="files => {
-            importSpecificFile(files)
-            mainMenuVisible = false
-          }">
-            <PopoverMenuItem>导入 pptist 文件</PopoverMenuItem>
-          </FileInput>
-          <FileInput accept="application/vnd.openxmlformats-officedocument.presentationml.presentation"  @change="files => {
-            importPPTXFile(files)
-            mainMenuVisible = false
-          }">
-            <PopoverMenuItem>导入 pptx 文件（测试版）</PopoverMenuItem>
-          </FileInput>
-          <PopoverMenuItem @click="setDialogForExport('pptx')">导出文件</PopoverMenuItem>
-          <PopoverMenuItem @click="resetSlides(); mainMenuVisible = false">重置幻灯片</PopoverMenuItem>
-          <PopoverMenuItem @click="goLink('https://github.com/pipipi-pikachu/PPTist/issues')">意见反馈</PopoverMenuItem>
-          <PopoverMenuItem @click="goLink('https://github.com/pipipi-pikachu/PPTist/blob/master/doc/Q&A.md')">常见问题</PopoverMenuItem>
-          <PopoverMenuItem @click="mainMenuVisible = false; hotkeyDrawerVisible = true">快捷键</PopoverMenuItem>
-        </template>
-        <div class="menu-item"><IconHamburgerButton class="icon" /></div>
-      </Popover>
+      <FileMenu></FileMenu>
 
       <div class="title">
         <Input
@@ -52,16 +31,6 @@
         <div class="menu-item"><IconGithub class="icon" /></div>
       </a>
     </div>
-
-    <Drawer
-      :width="320"
-      v-model:visible="hotkeyDrawerVisible"
-      placement="right"
-    >
-      <HotkeyDoc />
-    </Drawer>
-
-    <FullscreenSpin :loading="exporting" tip="正在导入..." />
   </div>
 </template>
 
@@ -69,29 +38,17 @@
 import { nextTick, ref } from 'vue'
 import { storeToRefs } from 'pinia'
 import { useMainStore, useSlidesStore } from '../../store'
-import useScreening from '../../hooks/useScreening'
-import useImport from '../../hooks/useImport'
-import useSlideHandler from '../../hooks/useSlideHandler'
 import type { DialogForExportTypes } from '../../types/export'
 
-import HotkeyDoc from './HotkeyDoc.vue'
-import FileInput from '../../components/FileInput.vue'
-import FullscreenSpin from '../../components/FullscreenSpin.vue'
-import Drawer from '../../components/Drawer.vue'
 import Input from '../../components/Input.vue'
-import Popover from '../../components/Popover.vue'
-import PopoverMenuItem from '../../components/PopoverMenuItem.vue'
+import FileMenu from '../FileMenu/FileMenu.vue'
 import EnterScreen from '../EnterScreen/EnterScreen.vue'
 
 const mainStore = useMainStore()
 const slidesStore = useSlidesStore()
 const { title } = storeToRefs(slidesStore)
-const { enterScreening, enterScreeningFromStart } = useScreening()
-const { importSpecificFile, importPPTXFile, exporting } = useImport()
-const { resetSlides } = useSlideHandler()
 
 const mainMenuVisible = ref(false)
-const hotkeyDrawerVisible = ref(false)
 const editingTitle = ref(false)
 const titleInputRef = ref<InstanceType<typeof Input>>()
 const titleValue = ref('')
@@ -105,11 +62,6 @@ const startEditTitle = () => {
 const handleUpdateTitle = () => {
   slidesStore.setTitle(titleValue.value)
   editingTitle.value = false
-}
-
-const goLink = (url: string) => {
-  window.open(url)
-  mainMenuVisible.value = false
 }
 
 const setDialogForExport = (type: DialogForExportTypes) => {
